Project 2: Monolith
In attempting to recreate the monolith from 2001: A Space Odyssey, I used brightness from video as an input and created two outputs: a projection of an animation determined by brightness, and the film's soundtrack, which plays when you get close to the monolith with a flashlight. A series of if statements set off a variety of sound effects that intensify as the brightness gets closer to the centre of the video input.